requirements.txt
streamlit==1.40.0
streamlit-drawable-canvas
Pillow
numpy
